+++ vers/changelog.md
@@ -6,7 +6,7 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
+* Adjusted some contract signatures to support block wise up/downloads  (**breaking Change**)